0.1: Controllers now read from global and local parameter files.
Fixed motor_controller2 to read all its parameters from the local
config file. Had to change some of the variables from static const
to just standard ones because it seems to be impossible to initialise
them by reading from the server.

Wrote some functions in motor_controller2 for generic parameter reading.
Should be moved to a utility directory somewhere globally accessible.
The functions read a parameter from the server, and error if it is not
present.